Add PianoBase.utils.ts & Refactored PianoBase to leave here only React Logic
