+++ app/pianobase/pianobase.tsx
@@ -2,90 +2,20 @@ import * as Tone from "tone";
 import React, { useState, useEffect, useRef } from "react";
 import { Button, DropdownMenu } from "@radix-ui/themes";
 import { HamburgerMenuIcon } from "@radix-ui/react-icons";
-import type { PianoBaseProps, chordMapType, OctaveRangeType } from "./PianoBase.types";
+import type { PianoBaseProps } from "./PianoBase.types";
+import {
+  defaultChordMap,
+  generateNotes,
+  getAlternativeNotation,
+  getBlackKeyLeft,
+  getBlackKeyWidth,
+  createDefaultSynth,
+  playNote as playNoteUtil,
+  playChord as playChordUtil,
+  playSequence as playSequenceUtil
+} from "./PianoBase.utils";
 
 import './PianoBase.css';
-
-const defaultChordMap: chordMapType = {
-  D4maj: ["D4", "A4", "F#5", "A5", "D6"],
-  E4min: ["E4", "B4", "G5", "B5", "E6"],
-  G4bmin: ["F#4", "C#5", "A5", "C#6", "F#6"],
-  G4maj: ["G4", "D5", "B5", "D6", "G6"],
-  A4maj: ["A4", "E5", "C#6", "E6", "A6"],
-  B4min: ["B4", "F#5", "D6", "F#6", "B6"],
-  C4dim: ["C#5", "G5", "E6", "G6", "C#7"],
-
-  D5maj: ["D5", "A5", "F#6", "A6", "D7"],
-};
-
-const generateNotes = function (octaves: OctaveRangeType = 3, startOctave: OctaveRangeType = 4) {
-  const white = [];
-  const black = [];
-
-  for (let i = 0; i < octaves; i++) {
-    const currentOctave = startOctave + i;
-    white.push(...["C", "D", "E", "F", "G", "A", "B"].map(n => `${n}${currentOctave}`));
-    black.push(...["C#", "D#", "F#", "G#", "A#"].map(n => `${n}${currentOctave}`));
-  }
-
-  // Agrega la nota final del último do (ej. C6)
-  white.push(`C${startOctave + octaves}`);
-
-  return { white, black };
-}
-
-function createDefaultSynth() {
-  // Sintetizador principal para el tono del piano
-  const synth = new Tone.PolySynth(Tone.Synth, {
-    volume: -8,
-    envelope: {
-      attack: 0.002,    // Ataque muy rápido para el golpe de martillo
-      decay: 0.5,       // Decay moderado
-      sustain: 0.15,    // Sustain bajo para simular las cuerdas del piano
-      release: 1.5      // Release largo para la resonancia natural
-    },
-    oscillator: {
-      type: "sine"      // Onda sinusoidal para un tono más puro
-    }
-  });
-
-  // Filtro para dar forma al sonido del piano
-  const filter = new Tone.Filter({
-    type: "lowpass",
-    frequency: 5000,    // Frecuencia de corte alta para mantener brillo
-    Q: 1               // Resonancia suave
-  });
-
-  // Compressor para controlar la dinámica
-  const compressor = new Tone.Compressor({
-    threshold: -20,
-    ratio: 3,
-    attack: 0.003,
-    release: 0.25
-  });
-
-  // Reverb sutil para simular la caja de resonancia del piano
-  const reverb = new Tone.Reverb({
-    decay: 1.5,        // Decay moderado
-    wet: 0.2          // Mezcla sutil
-  }).toDestination();
-
-  // Conectamos la cadena de efectos
-  synth.chain(filter, compressor, reverb);
-
-  // Retornamos un objeto compatible con la interfaz esperada
-  return {
-    triggerAttackRelease(note: string, duration: string | number) {
-      synth.triggerAttackRelease(note, duration);
-    },
-    dispose() {
-      synth.dispose();
-      filter.dispose();
-      compressor.dispose();
-      reverb.dispose();
-    }
-  } as unknown as Tone.PolySynth;
-}
 
 export default function PianoBase({
   createSynth,
@@ -100,28 +30,12 @@ export default function PianoBase({
   const [showChords, setShowChords] = useState(false);
   const { white, black } = generateNotes(octaves, octave);
 
-  const sharpToFlatMap: Record<string, string> = {
-    "C#": "Db",
-    "D#": "Eb",
-    "F#": "Gb",
-    "G#": "Ab",
-    "A#": "Bb"
-  };
-
-  const getAlternativeNotation = (note: string): string => {
-    const [, noteName, octave] = note.match(/([A-G]#)(\d)/) || [];
-    if (noteName && sharpToFlatMap[noteName]) {
-      return `${sharpToFlatMap[noteName]}${octave}`;
-    }
-    return "";
-  };
-
   useEffect(() => {
     const synth = createSynth
       ? createSynth()
       : createDefaultSynth();
 
-    synthRef.current = synth;  // Asignar el sintetizador al ref
+    synthRef.current = synth;
 
     return () => {
       synth.dispose();
@@ -132,23 +46,23 @@ export default function PianoBase({
   useEffect(() => {
     if (showChordOnThePiano && showChordOnThePiano.length > 0) {
       setHighlightedKeys(showChordOnThePiano);
-      playChord(showChordOnThePiano);
+      playChordUtil(showChordOnThePiano, synthRef.current);
     }
   }, [showChordOnThePiano]);
 
-  const playNote = (note: string) => {
+  const handlePlayNote = (note: string) => {
     setActiveNotes([note]);
-    synthRef.current?.triggerAttackRelease(note, "4n");
+    playNoteUtil(note, synthRef.current);
     setTimeout(() => setActiveNotes([]), 180);
   };
 
-  const playChord = (notes: string[]) => {
+  const handlePlayChord = (notes: string[]) => {
     setActiveNotes(notes);
-    notes.forEach(note => synthRef.current?.triggerAttackRelease(note, "2n"));
+    playChordUtil(notes, synthRef.current);
     setTimeout(() => setActiveNotes([]), 180);
   };
 
-  const playSequence = (notesOrName: string[] | string, delay = 200) => {
+  const handlePlaySequence = (notesOrName: string[] | string) => {
     const notes = Array.isArray(notesOrName)
       ? notesOrName
       : chordMap[notesOrName] || [];
@@ -156,47 +70,9 @@ export default function PianoBase({
     if (notes.length === 0) return;
 
     setActiveNotes(notes);
-
-    notes.forEach((note, i) => {
-      setTimeout(() => {
-        synthRef.current?.triggerAttackRelease(note, "8n");
-        setActiveNotes([note]);
-      }, i * delay);
-    });
-    setTimeout(() => setActiveNotes([]), notes.length * delay + 180);
+    playSequenceUtil(notesOrName, synthRef.current, chordMap);
+    setTimeout(() => setActiveNotes([]), notes.length * 200 + 180);
   };
-
-  const getBlackKeyLeft = (note: string, whiteNotes: string[]) => {
-    const blackToWhiteBefore: Record<string, string> = {
-      "C#": "C",
-      "D#": "D",
-      "F#": "F",
-      "G#": "G",
-      "A#": "A",
-    };
-
-    const match = note.match(/^([A-G]#)(\d)$/);
-    if (!match) return "0%";
-
-    const [_, pitchClass, octave] = match;
-    const whiteBefore = `${blackToWhiteBefore[pitchClass]}${octave}`;
-    const whiteIndex = whiteNotes.indexOf(whiteBefore);
-
-    if (whiteIndex === -1) return "0%";
-
-    const whiteKeyWidth = 100 / whiteNotes.length;
-    const left = (whiteIndex + 1) * whiteKeyWidth;
-
-    return `${left}%`; // ya está centrado por transform: translateX(-50%)
-  };
-
-  function getBlackKeyWidth(octaves: OctaveRangeType): string {
-    if (octaves <= 1) return "7%";
-    if (octaves === 2) return "4%";
-    if (octaves === 3) return "3%";
-    if (octaves === 4) return "2%";
-    return "1.4%";
-  }
 
   return (
     <div className="piano-base">
@@ -214,7 +90,7 @@ export default function PianoBase({
               {Object.entries(chordMap).map(([chordName], i) => (
                 <DropdownMenu.Item
                   key={chordName}
-                  onClick={() => playSequence(chordName)}
+                  onClick={() => handlePlaySequence(chordName)}
                   color={["orange", "yellow", "green", "blue", "indigo", "purple", "red"][i % 7]}
                 >
                   Play {chordName}
@@ -222,14 +98,13 @@ export default function PianoBase({
               ))}
             </DropdownMenu.SubContent>
           </DropdownMenu.Sub>
-
         </DropdownMenu.Content>
       </DropdownMenu.Root>)}
 
       {showChords && Object.entries(chordMap).map(([chordName], i) => (
         <Button
           key={chordName}
-          onClick={() => playSequence(chordName)}
+          onClick={() => handlePlaySequence(chordName)}
           variant="classic"
           color={["orange", "yellow", "green", "blue", "indigo", "purple", "red"][i % 7]}
         >
@@ -244,7 +119,7 @@ export default function PianoBase({
               key={note}
               className={`white-key${activeNotes.includes(note) || highlightedKeys.includes(note) ? " active-key" : ""}`}
               data-note={note}
-              onClick={() => playNote(note)}
+              onClick={() => handlePlayNote(note)}
             >{highlightedKeys.includes(note) && <span className="note-name">{note}</span>}</div>
           ))}
         </div>
@@ -259,7 +134,7 @@ export default function PianoBase({
                 width: getBlackKeyWidth(octaves)
               }}
               data-note={note}
-              onClick={() => playNote(note)}
+              onClick={() => handlePlayNote(note)}
             >{highlightedKeys.includes(note) && <span className="note-name">
               <span className="flat-notation">{getAlternativeNotation(note)}</span>
               <span className="sharp-notation">{note}</span>
